+++ NEWS.md
@@ -16,6 +16,8 @@
 * Added support for `dplyr::lead()`. In `dplyr::lead()` and `dplyr::lag()`, the 
   arguments `default` and `order_by` are now supported (#151).
 
+* Added support for `stringr::str_replace_na()` (#153)
+
 # tidypolars 0.11.0
 
 `tidypolars` requires `polars` >= 0.20.0.